+++ DBProject/BBScout/dbo/Tables/TeamAttribute.sql
@@ -1,15 +1,17 @@
 ﻿CREATE TABLE [dbo].[TeamAttribute] (
-    [id]           INT            IDENTITY (1, 1) NOT NULL,
-    [teamId]       INT            NOT NULL,
-    [attributeId]  INT            NOT NULL,
-    [integerValue] INT            NULL,
-    [decimalValue] INT            NULL,
-    [textValue]    VARCHAR (4000) NULL,
-    [lastUpdated]  DATETIME       NULL,
+    [id]           INT             IDENTITY (1, 1) NOT NULL,
+    [teamId]       INT             NOT NULL,
+    [attributeId]  INT             NOT NULL,
+    [integerValue] INT             NULL,
+    [decimalValue] DECIMAL (18, 2) NULL,
+    [textValue]    VARCHAR (4000)  NULL,
+    [lastUpdated]  DATETIME        NULL,
     PRIMARY KEY CLUSTERED ([id] ASC),
     CONSTRAINT [fk_TeamAttribute_Attribute] FOREIGN KEY ([attributeId]) REFERENCES [dbo].[Attribute] ([id]),
     CONSTRAINT [fk_TeamAttribute_Team] FOREIGN KEY ([teamId]) REFERENCES [dbo].[Team] ([id])
 );
+
+
 
 
 GO
